move functions to a separate file
